% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate-water-temp-data.R
\name{aggregate_water_temp_data}
\alias{aggregate_water_temp_data}
\title{Aggregate Water Temperature Data}
\usage{
aggregate_water_temp_data(
  data,
  ...,
  min_coverage = 0.875,
  date_time = "date_time",
  value = "water_temperature"
)
}
\arguments{
\item{data}{A data frame.}

\item{...}{These dots are for future extensions and must be empty.}

\item{min_coverage}{A numeric value of the minimum coverage as a proportion.}

\item{date_time}{A string indicating the column name of the POSIXct vector.}

\item{value}{A string indicating the column name of the value vector.}
}
\value{
A data frame with a date column and update water temperature column.
}
\description{
Aggregates subdaily water temperature data to mean daily water temperature.
If the water temperature data spans less than the min_coverage than
a missing value is return instead.
}
\details{
The min_coverage is converted into the minimum number of non-missing
values in each day based on how many values of the shortest time interval
within the time series are required to achieve the minimum coverage.
For example if the date times are every 15 minutes then 24 / 0.25 = 96 values
are required for 100\% coverage and 84 values are required for the default
87.5\% minimum coverage.
}
\examples{
data <- data.frame(
  date_time =
    as.POSIXct(c(
      "2021-05-07 00:00:00", "2021-05-07 08:00:00",
      "2021-05-07 16:00:00"
    )),
  water_temperature = c(5, 5, 7)
)

aggregate_water_temp_data(data)
}
