% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evrfish-package.R
\docType{package}
\name{evrfish-package}
\alias{evrfish}
\alias{evrfish-package}
\title{evrfish: R Tools for EVR Fish Projects}
\description{
Provides functions linearly interpolate missing values and calculate growing season degree days and growing degree days from daily water temperature data. It also exports fish codes for British Columbia and Alberta from the `fishbc` package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://poissonconsulting.github.io/evrfish/}
  \item \url{https://github.com/poissonconsulting/evrfish}
  \item Report bugs at \url{https://github.com/poissonconsulting/evrfish/issues/}
}

}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-7683-4592}{ORCID})

Authors:
\itemize{
  \item Joe Thorley \email{joe@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-7683-4592}{ORCID})
  \item Ayla Pearson \email{ayla@poissonconsulting.ca} (\href{https://orcid.org/0000-0001-7388-1222}{ORCID})
  \item Sarah Lyons (\href{https://orcid.org/0000-0002-6745-6796}{ORCID})
}

Other contributors:
\itemize{
  \item Bronwen Lewis [contributor]
  \item Jill Brooks [contributor]
  \item Andrew Harwood [contributor]
  \item Sebastian Dalgarno (\href{https://orcid.org/0000-0002-3658-4517}{ORCID}) [contributor]
  \item Elk Valley Resources [funder, copyright holder]
}

}
\keyword{internal}
