% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{freshwaterfish}
\alias{freshwaterfish}
\title{BC Fish Data}
\usage{
freshwaterfish
}
\description{
Curated data on the codes,
classification and conservation status of freshwater fishes in British Columbia.
}
