% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdd.R
\name{gdd}
\alias{gdd}
\title{Calculate Growing Degree Days (GDD)}
\usage{
gdd(x, end_date = as.Date("1972-09-30"), min_length = 60, msgs = TRUE)
}
\arguments{
\item{x}{A data frame with two columns \code{date} and \code{temperature}.
\code{date}, which must be of class Date provides the dates and
\code{temperature} which must be a numeric vector provides the
mean daily water temperature in degrees centigrade.}

\item{end_date}{A Date scalar of the last date
within each year to consider (the year is ignored).}

\item{min_length}{A whole number of the minimum number of values to consider.}

\item{msgs}{A flag specifying whether to provide messages.}
}
\description{
A wrapper on \code{\link[gsdd:gdd]{gsdd::gdd()}} to get the Growing Degree Days up to a date for
all growing seasons.
}
\examples{
gdd(gsdd::temperature_data)
}
\seealso{
\code{\link[gsdd:gdd]{gsdd::gdd()}}, \code{\link[=gsdd]{gsdd()}} and \code{\link[=gss]{gss()}}.
}
