% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsdd.R
\name{gsdd}
\alias{gsdd}
\title{Calculate Growing Season Degree Days (GSDD)}
\usage{
gsdd(x, min_length = 120, msgs = TRUE)
}
\arguments{
\item{x}{A data frame with two columns \code{date} and \code{temperature}.
\code{date}, which must be of class Date provides the dates and
\code{temperature} which must be a numeric vector provides the
mean daily water temperature in degrees centigrade.}

\item{min_length}{A whole number of the minimum number of values to consider.}

\item{msgs}{A flag specifying whether to provide messages.}
}
\description{
A wrapper on \code{\link[gsdd:gsdd]{gsdd::gsdd()}} to get the Growing Season Degree Days for
all the growing seasons in a year.
}
\examples{
gsdd(gsdd::temperature_data)
}
\seealso{
\code{\link[gsdd:gsdd]{gsdd::gsdd()}}, \code{\link[=gsdd]{gsdd()}} and \code{\link[=gss]{gss()}}.
}
