% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsdd-cf.R
\name{gsdd_cf}
\alias{gsdd_cf}
\title{Calculate Growing Season Degree Days (GSDD)}
\usage{
gsdd_cf(x, ignore_truncation = FALSE, min_length = 120, msgs = TRUE)
}
\arguments{
\item{x}{A numeric vector of the
mean daily water temperature values for the period
of interest in C.}

\item{ignore_truncation}{A flag specifying whether to ignore truncation
of the mean daily water temperature vector
or a string of "start", "end", "none" (equivalent to FALSE) or "both"
(equivalent to TRUE) specifying which type of truncation to ignore.}

\item{min_length}{A whole number of the minimum number of values to consider.}

\item{msgs}{A flag specifying whether to provide messages.}
}
\value{
A non-negative real number of the GSDD.
}
\description{
Soft-deprecated for \code{\link[gsdd:gsdd_vctr]{gsdd::gsdd_vctr()}}.
}
\examples{
gsdd_cf(c(rep(1, 10), rep(10, 20), rep(1, 200)))
gsdd_cf(gsdd::temperature_data$temperature)
}
\seealso{
\code{\link[gsdd:gsdd_vctr]{gsdd_vctr()}}
}
