% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss.R
\name{gss}
\alias{gss}
\title{Calculate Growing Seasons (GSS)}
\usage{
gss(x, min_length = 120, ignore_truncation = "end", pick = "all", msgs = TRUE)
}
\arguments{
\item{x}{A data frame with two columns \code{date} and \code{temperature}.
\code{date}, which must be of class Date provides the dates and
\code{temperature} which must be a numeric vector provides the
mean daily water temperature in degrees centigrade.}

\item{min_length}{A whole number of the minimum number of values to consider.}

\item{ignore_truncation}{A flag specifying whether to ignore truncation
of the mean daily water temperature vector
or a string of "start", "end", "none" (equivalent to FALSE) or "both"
(equivalent to TRUE) specifying which type of truncation to ignore.}

\item{pick}{A string specifying which 'season' to pick when more than one
is present. One of "longest", "shortest", "first" or "last", or the season
with the "biggest" or "smallest" GSDD, or "all" to use every season.
By default "all".}

\item{msgs}{A flag specifying whether to provide messages.}
}
\description{
A wrapper on \code{\link[gsdd:gss]{gsdd::gss()}} to by default
get all Growing Seasons ignoring truncation.
For more information see \code{\link[gsdd:gss]{gsdd::gss()}}.
}
\examples{
gss(gsdd::temperature_data)
}
\seealso{
\code{\link[gsdd:gss]{gsdd::gss()}}, \code{\link[=gsdd]{gsdd()}} and \code{\link[=gss]{gss()}}.
}
