% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss-plot.R
\name{gss_plot}
\alias{gss_plot}
\title{Plot Growing Seasons (GSS)}
\usage{
gss_plot(
  x,
  min_length = 60,
  ignore_truncation = TRUE,
  pick = "all",
  latex = FALSE,
  nrow = NULL,
  ncol = NULL,
  msgs = TRUE
)
}
\arguments{
\item{x}{A data frame with two columns \code{date} and \code{temperature}.
\code{date}, which must be of class Date provides the dates and
\code{temperature} which must be a numeric vector provides the
mean daily water temperature in degrees centigrade.}

\item{min_length}{A whole number of the minimum number of values to consider.}

\item{ignore_truncation}{A flag specifying whether to ignore truncation
of the mean daily water temperature vector
or a string of "start", "end", "none" (equivalent to FALSE) or "both"
(equivalent to TRUE) specifying which type of truncation to ignore.}

\item{pick}{A string specifying which 'season' to pick when more than one
is present. One of "longest", "shortest", "first" or "last", or the season
with the "biggest" or "smallest" GSDD, or "all" to use every season.
By default "all".}

\item{latex}{A flag specifying whether to use LaTeX to include degree
symbol in y-axis label.}

\item{nrow}{A count of the number of rows to facet by.}

\item{ncol}{A count of the number of columns to facet by.}

\item{msgs}{A flag specifying whether to provide messages.}
}
\description{
A wrapper on \code{\link[gsdd:gss_plot]{gsdd::gss_plot()}} to by default
plot all Growing Seasons ignoring truncation.
For more information see \code{\link[gsdd:gss_plot]{gsdd::gss_plot()}}.
}
\examples{
gss_plot(gsdd::temperature_data)
}
\seealso{
\code{\link[gsdd:gss_plot]{gsdd::gss_plot()}} and \code{\link[=gss]{gss()}}.
}
