% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate-numeric-vector.R
\name{interpolate_numeric_vector}
\alias{interpolate_numeric_vector}
\title{Interpolate Numeric Vector}
\usage{
interpolate_numeric_vector(x, span = 3, tails = FALSE)
}
\arguments{
\item{x}{A double or integer vector of with missing values to fill
in using linear interpolation.}

\item{span}{A whole number of the maximum span of missing values to interpolate.
If a gap exceeds the span none of the values are interpolate.}

\item{tails}{A flag specifying whether to fill in missing values at the
start and end by setting them to be the same value as the closest
adjacent non-missing value.}
}
\value{
A double or integer vector.
}
\description{
Useful for filling in short runs of missing values in a time series.
}
\details{
\code{interpolate_numeric_vector()} is essentially a wrapper on \code{\link[stats:approx]{stats::approx()}}.
}
\examples{
interpolate_numeric_vector(c(1, NA, 4))
interpolate_numeric_vector(c(1L, NA, 4L))
interpolate_numeric_vector(c(1, NA, NA, NA, NA, 3))
interpolate_numeric_vector(c(1, NA, NA, NA, NA, 3), span = 4)
interpolate_numeric_vector(c(NA, NA, 10, 1, NA))
interpolate_numeric_vector(c(NA, NA, 10, 1, NA), tails = TRUE)
}
