% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish-codes.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{fbc_common_name}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{fishbc}{\code{\link[fishbc:fbc_common_name]{fbc_common_name()}}}
}}

